# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: Authenticator API
    version: 1.0.0
    description: |
        REST API for user identity, login and OpenID Connect authentication.
servers:
    - url: /

tags:
    - name: Health
      description: Server health check
    - name: Users
      description: User and authentication management

paths:
    /health:
        get:
            tags:
                - Health
            summary: Health check
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string

    /api/v1/users:
        get:
            tags:
                - Users
            summary: List all users
            operationId: getUsers
            security:
                - BearerAuth: []
            description: |
                Returns an array with all users in the system.
                Requires admin role to access.
            responses:
                '200':
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/User'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/InternalError'
        post:
            tags:
                - Users
            summary: Create a new user
            operationId: createUser
            security:
                - BearerAuth: []
            description: |
                Creates a new user with the provided credentials.
                Requires admin role to create users.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserInput'
            responses:
                '201':
                    description: User created successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '409':
                    $ref: '#/components/responses/Conflict'
                '500':
                    $ref: '#/components/responses/InternalError'

    /api/v1/users/username/{username}:
        parameters:
            - in: path
              name: username
              required: true
              schema:
                  type: string
              description: Username of the user
              example: johndoe
        get:
            tags:
                - Users
            summary: Get user by username
            operationId: getUserByUsername
            security:
                - BearerAuth: []
            description: |
                Returns a specific user by their username.
                Requires admin role to access.
            responses:
                '200':
                    description: User found
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            tags:
                - Users
            summary: Update user by username
            operationId: updateUserByUsername
            security:
                - BearerAuth: []
            description: |
                Updates an existing user's information by username.
                Requires admin role to update users.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserInput'
            responses:
                '200':
                    description: User updated successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Users
            summary: Delete user by username
            operationId: deleteUserByUsername
            security:
                - BearerAuth: []
            description: |
                Deletes a user from the system by username.
                Requires admin role to delete users.
            responses:
                '200':
                    description: User deleted successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              nullable: true
                                              example: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    /api/v1/users/{id}:
        parameters:
            - in: path
              name: id
              required: true
              schema:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
              description: MongoDB ObjectId of the user
              example: 665c2d3e4f5a6b7c8d9e0f1a
        get:
            tags:
                - Users
            summary: Get user by id
            operationId: getUserById
            security:
                - BearerAuth: []
            description: |
                Returns a specific user by their MongoDB ObjectId.
                Requires admin role to access.
            responses:
                '200':
                    description: User found
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            tags:
                - Users
            summary: Update user by id
            operationId: updateUserById
            security:
                - BearerAuth: []
            description: |
                Updates an existing user's information by MongoDB ObjectId.
                Requires admin role to update users.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserInput'
            responses:
                '200':
                    description: User updated successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/User'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Users
            summary: Delete user by id
            operationId: deleteUserById
            security:
                - BearerAuth: []
            description: |
                Deletes a user from the system by MongoDB ObjectId.
                Requires admin role to delete users.
            responses:
                '200':
                    description: User deleted successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              nullable: true
                                              example: null
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    /api/v1/users/login:
        post:
            tags:
                - Users
            summary: User login
            operationId: login
            description: |
                Authenticates a user using their login credentials (username or email) and password.
                Returns a JWT token upon successful authentication.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginInput'
            responses:
                '200':
                    description: Login successful. Returns a JWT token.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  token:
                                                      type: string
                                                      description: JWT authentication token
                                                      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '500':
                    $ref: '#/components/responses/InternalError'

    /api/v1/users/oidc/login:
        post:
            tags:
                - Users
            summary: Start OIDC login
            operationId: oidcLogin
            description: |
                Builds the OIDC provider authorization URL.
                Requires OIDC to be enabled in the server configuration.
            responses:
                '200':
                    description: OIDC login URL generated successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  loginUrl:
                                                      type: string
                                                      format: uri
                                                      description: URL where the client should redirect the user
                                                      example: https://identity.example.com/authorize?client_id=scope-manager&response_type=code
                '400':
                    $ref: '#/components/responses/ValidationError'
                '500':
                    $ref: '#/components/responses/InternalError'

    /api/v1/users/oidc/callback:
        get:
            tags:
                - Users
            summary: Complete OIDC login
            operationId: oidcCallback
            description: |
                Handles the OIDC provider callback and returns a JWT token for the associated local user.
                Requires OIDC to be enabled in the server configuration.
            parameters:
                - in: query
                  name: code
                  required: true
                  schema:
                      type: string
                  description: Authorization code returned by the OIDC provider
                - in: query
                  name: state
                  required: false
                  schema:
                      type: string
                  description: OIDC state value returned by the provider, when present
            responses:
                '200':
                    description: OIDC login completed successfully. Returns a JWT token.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  token:
                                                      type: string
                                                      description: JWT authentication token
                                                      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '500':
                    $ref: '#/components/responses/InternalError'

components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT token obtained via POST /api/v1/users/login

    responses:
        ValidationError:
            description: Validation error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Unauthorized:
            description: Missing or invalid authentication token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Forbidden:
            description: Insufficient permissions to perform this action
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        NotFound:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Conflict:
            description: Resource conflict
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'

    schemas:
        SuccessResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                data:
                    nullable: true
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error

        ErrorResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: false
                message:
                    type: string
                    example: Validation failed
                httpStatus:
                    type: integer
                    example: 400
                appCode:
                    type: string
                    example: VALIDATION_ERROR
                data:
                    type: object
                    nullable: true
                    example: null
                error:
                    type: object
                    nullable: true
                    properties:
                        message:
                            type: string
                        details:
                            nullable: true

        User:
            type: object
            properties:
                _id:
                    type: string
                    description: MongoDB ObjectId
                    example: 665c2d3e4f5a6b7c8d9e0f1a
                username:
                    type: string
                    example: johndoe
                email:
                    type: string
                    format: email
                    example: john@example.com
                name:
                    type: string
                    example: John
                surname:
                    type: string
                    example: Doe
                systemRole:
                    type: string
                    enum: [ADMIN, USER]
                    example: ADMIN
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time

        UserInput:
            type: object
            required:
                - username
                - email
                - password
                - name
                - surname
                - systemRole
            properties:
                username:
                    type: string
                    example: johndoe
                email:
                    type: string
                    format: email
                    example: john@example.com
                password:
                    type: string
                    format: password
                    example: securePassword123
                name:
                    type: string
                    example: John
                surname:
                    type: string
                    example: Doe
                systemRole:
                    type: string
                    enum: [ADMIN, USER]
                    example: ADMIN

        LoginInput:
            type: object
            required:
                - login
                - password
            properties:
                login:
                    type: string
                    description: Username or email
                    example: johndoe
                password:
                    type: string
                    format: password
                    example: securePassword123